requirements.txt
streamlit==1.26.0
pandas==2.1.1
numpy==1.23.5
Pillow==9.5.0
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.8.0
ultralytics==8.0.158
filterpy==1.4.5
scikit-learn==1.3.0
